+++ apps/learning-snippets/src/test/property-grid/Telemetry.test.tsx
@@ -3,6 +3,7 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 /* eslint-disable import/no-duplicates */
+/* eslint-disable deprecation/deprecation */
 
 import { expect } from "chai";
 import sinon from "sinon";