+++ apps/test-viewer/src/UiProvidersConfig.tsx
@@ -226,7 +226,16 @@ const configuredUiItems = new Map<string, UiItem>([
         await MeasureTools.startup();
         MeasurementActionToolbar.setDefaultActionProvider();
       },
-      createUiItemsProviders: () => [new MeasureToolsUiItemsProvider()],
+      createUiItemsProviders: () => [new MeasureToolsUiItemsProvider({
+        measurementFormattingProps: {
+          distance: {
+            bearing: {
+              koqName: "RoadRailUnits.Bearing",
+              persistenceUnitName: "Units.RAD"
+            }
+          }
+        }
+      })],
     },
   ],
   [
@@ -245,6 +254,8 @@ const configuredUiItems = new Map<string, UiItem>([
       initialize: async () => {
         await GeoTools.initialize();
       },
+      // TODO: Use next line to use Google API
+      // createUiItemsProviders: () => [new GeoToolsAddressSearchProvider(new GoogleAddressProvider)],
       createUiItemsProviders: () => [new GeoToolsAddressSearchProvider()],
     },
   ],