+++ packages/itwin/measure-tools/eslint.config.js
@@ -15,7 +15,6 @@ export default [
           tag: [
             "internal",
             "alpha",
-            "beta",
           ],
         },
       ],